+++ src/pages/api/plays/[id].ts
@@ -1,6 +1,7 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
-import type { Prisma } from "@prisma/client"
+import { Prisma, Role } from "@prisma/client"
 import models from 'lib/models';
+import { getSession } from 'next-auth/react';
 
 const getONE = async (playId: number, response: NextApiResponse) => {
   try {
@@ -74,6 +75,7 @@ const updateONE = async (body: any, playId: number, response: NextApiResponse) =
 const doRequest = async (request: NextApiRequest, response: NextApiResponse) => {
   const { method, query, body } = request;
   const playId: number = parseInt(query.id as string, 10);
+  const session = await getSession({req: request})
 
   if (method === 'GET') {
     getONE(playId, response)
@@ -81,13 +83,18 @@ const doRequest = async (request: NextApiRequest, response: NextApiResponse) => 
     return
   }
 
-  if (method === 'DELETE') {
+  if (!session) {
+    response.status(401).json({err: "unauthorized"});
+    return 
+  }
+
+  if (method === 'DELETE' && session.user.role === Role.ADMIN) {
     deleteONE(playId, response)
 
     return
   }
 
-  if (method === 'PUT') {
+  if (method === 'PUT' && session.user.role === Role.ADMIN) {
     updateONE(body, playId, response)
 
     return
